refactor(hooks): relocate tier classifier to core/hooks + refspec + dominant_lens
Stage 4b: move core/practice/irreversible_tier.py to
core/hooks/_irreversible_tier.py (underscore hook-private convention,
matching _scenario_detector / _cascade_detector / _blueprint_d). The old
path becomes a sys.modules-aliasing compat shim so both import paths
resolve to ONE module object (single Tier enum identity). All import
sites repointed (reasoning_surface_guard lazy import, _tier1_audit, the
4 test files); the installed-CLI sys.path guard repoints from
core/practice to the stable core/ package-root invariant.

Adds refspec normalization (_normalize_branch_ref handles refs/heads/x
and HEAD:x) to validate_micro_surface so legitimate refspec pushes are
not false-flagged as branch-binding mismatches; unified into
_resolve_push_target. Adds the additive, inert OperatorProfile.dominant_lens
axis + both disk loaders (json list + markdown value:[a,b,c]).

scenario_detector audited: it does NOT over-fire on additive multi-file
work (the apparent noise was the separately-fixed non-dedup writer), so
no discriminator added. 116 tier tests green.
